+++ src/core/draw/frame/Margin.ts
@@ -14,12 +14,11 @@ export class Margin {
 
   public render(ctx2d: Context2d, pageNo: number) {
     const { marginIndicatorColor, pageMode } = this.options
-    const width = this.draw.getWidth()
+    const { width, height: pageHeight, margins } = this.draw.getPageSize(pageNo)
     const height =
       pageMode === PageMode.CONTINUITY
         ? this.draw.getCanvasHeight(pageNo) / this.draw.getPagePixelRatio()
-        : this.draw.getHeight()
-    const margins = this.draw.getMargins()
+        : pageHeight
     const marginIndicatorSize = this.draw.getMarginIndicatorSize()
     ctx2d.save()
     ctx2d.translate(0.5, 0.5)