+++ src/core/draw/frame/PageBorder.ts
@@ -18,7 +18,7 @@ export class PageBorder {
     this.options = draw.getOptions()
   }
 
-  public render(ctx2d: Context2d) {
+  public render(ctx2d: Context2d, pageNo: number) {
     const {
       scale,
       pageBorder: { color, lineWidth, padding }
@@ -27,18 +27,23 @@ export class PageBorder {
     ctx2d.translate(0.5, 0.5)
     ctx2d.strokeStyle = color
     ctx2d.lineWidth = lineWidth * scale
-    const margins = this.draw.getMargins()
+    const {
+      margins,
+      innerWidth,
+      height: pageHeight
+    } = this.draw.getPageSize(pageNo)
     // x：左边距 - 左距离正文距离
     const x = margins[3] - padding[3] * scale
     // y：页眉上边距 + 页眉高度 - 上距离正文距离
-    const y = margins[0] + this.header.getExtraHeight() - padding[0] * scale
+    const y =
+      margins[0] + this.header.getExtraHeight(pageNo) - padding[0] * scale
     // width：页面宽度 + 左右距离正文距离
-    const width = this.draw.getInnerWidth() + (padding[1] + padding[3]) * scale
+    const width = innerWidth + (padding[1] + padding[3]) * scale
     // height：页面高度 - 正文起始位置 - 页脚高度 - 下边距 - 下距离正文距离
     const height =
-      this.draw.getHeight() -
+      pageHeight -
       y -
-      this.footer.getExtraHeight() -
+      this.footer.getExtraHeight(pageNo) -
       margins[2] +
       padding[2] * scale
     ctx2d.rect(x, y, width, height)