+++ src/dataset/constant/Element.ts
@@ -85,7 +85,8 @@ export const EDITOR_ELEMENT_ZIP_ATTR: Array<keyof IElement> = [
   'hide',
   'label',
   'labelId',
-  'lineWidth'
+  'lineWidth',
+  'paperDirection'
 ]
 
 export const TABLE_TD_ZIP_ATTR: Array<keyof ITd> = [